[L1Trigger] Move public headers in to interface directory
