+++ L1Trigger/DTTrackFinder/src/L1MuDTDataBuffer.cc
@@ -31,8 +31,8 @@
 
 #include "L1Trigger/DTTrackFinder/interface/L1MuDTTFConfig.h"
 #include "L1Trigger/DTTrackFinder/src/L1MuDTSectorProcessor.h"
-#include "L1Trigger/DTTrackFinder/src/L1MuDTTrackSegLoc.h"
-#include "L1Trigger/DTTrackFinder/src/L1MuDTTrackSegPhi.h"
+#include "L1Trigger/DTTrackFinder/interface/L1MuDTTrackSegLoc.h"
+#include "L1Trigger/DTTrackFinder/interface/L1MuDTTrackSegPhi.h"
 
 using namespace std;
