+++ L1Trigger/DTTrackFinder/src/L1MuDTERS.cc
@@ -29,7 +29,7 @@
 //-------------------------------
 
 #include "L1Trigger/DTTrackFinder/interface/L1MuDTTFConfig.h"
-#include "L1Trigger/DTTrackFinder/src/L1MuDTTrackSegPhi.h"
+#include "L1Trigger/DTTrackFinder/interface/L1MuDTTrackSegPhi.h"
 #include "L1Trigger/DTTrackFinder/src/L1MuDTSEU.h"
 #include "L1Trigger/DTTrackFinder/src/L1MuDTEUX.h"
