+++ L1Trigger/DTTrackFinder/src/L1MuDTEtaProcessor.cc
@@ -35,8 +35,8 @@
 //-------------------------------
 
 #include "L1Trigger/DTTrackFinder/interface/L1MuDTTFConfig.h"
-#include "L1Trigger/DTTrackFinder/src/L1MuDTTrackSegEta.h"
-#include "L1Trigger/DTTrackFinder/src/L1MuDTSecProcId.h"
+#include "L1Trigger/DTTrackFinder/interface/L1MuDTTrackSegEta.h"
+#include "L1Trigger/DTTrackFinder/interface/L1MuDTSecProcId.h"
 #include "L1Trigger/DTTrackFinder/src/L1MuDTSectorProcessor.h"
 #include "L1Trigger/DTTrackFinder/interface/L1MuDTTrackFinder.h"
 #include "L1Trigger/DTTrackFinder/interface/L1MuDTTrack.h"
