+++ L1Trigger/DTTrackFinder/src/L1MuDTSecProcMap.h
@@ -28,7 +28,7 @@
 // Collaborating Class Declarations --
 //------------------------------------
 
-#include "L1Trigger/DTTrackFinder/src/L1MuDTSecProcId.h"
+#include "L1Trigger/DTTrackFinder/interface/L1MuDTSecProcId.h"
 class L1MuDTSectorProcessor;
 
 //              ---------------------
