+++ L1Trigger/DTTrackFinder/src/L1MuDTSectorProcessor.cc
@@ -31,7 +31,7 @@
 
 #include <FWCore/Framework/interface/Event.h>
 #include "L1Trigger/DTTrackFinder/interface/L1MuDTTFConfig.h"
-#include "L1Trigger/DTTrackFinder/src/L1MuDTSecProcId.h"
+#include "L1Trigger/DTTrackFinder/interface/L1MuDTSecProcId.h"
 #include "L1Trigger/DTTrackFinder/src/L1MuDTSectorReceiver.h"
 #include "L1Trigger/DTTrackFinder/src/L1MuDTDataBuffer.h"
 #include "L1Trigger/DTTrackFinder/src/L1MuDTExtrapolationUnit.h"
