+++ L1Trigger/DTTrackFinder/src/L1MuDTSectorProcessor.h
@@ -30,7 +30,7 @@
 // Base Class Headers --
 //----------------------
 
-#include "L1Trigger/DTTrackFinder/src/L1MuDTSecProcId.h"
+#include "L1Trigger/DTTrackFinder/interface/L1MuDTSecProcId.h"
 
 //------------------------------------
 // Collaborating Class Declarations --
