+++ L1Trigger/DTTrackFinder/src/L1MuDTTrackSegLoc.cc
@@ -15,7 +15,7 @@
 // This Class's Header --
 //-----------------------
 
-#include "L1Trigger/DTTrackFinder/src/L1MuDTTrackSegLoc.h"
+#include "L1Trigger/DTTrackFinder/interface/L1MuDTTrackSegLoc.h"
 
 //---------------
 // C++ Headers --
