+++ L1Trigger/GlobalCaloTrigger/src/L1GctJetCount.cc
@@ -1,5 +1,5 @@
 
-#include "L1Trigger/GlobalCaloTrigger/src/L1GctJetCount.h"
+#include "L1Trigger/GlobalCaloTrigger/interface/L1GctJetCount.h"
 
 // no code here, these are template classes
 // see http://www.parashift.com/c++-faq-lite/templates.html#faq-35.12
