+++ L1Trigger/GlobalCaloTrigger/test/testEtTypes.cpp
@@ -1,8 +1,8 @@
 
 
-#include "L1Trigger/GlobalCaloTrigger/src/L1GctTwosComplement.h"
-#include "L1Trigger/GlobalCaloTrigger/src/L1GctUnsignedInt.h"
-#include "L1Trigger/GlobalCaloTrigger/src/L1GctJetCount.h"
+#include "L1Trigger/GlobalCaloTrigger/interface/L1GctTwosComplement.h"
+#include "L1Trigger/GlobalCaloTrigger/interface/L1GctUnsignedInt.h"
+#include "L1Trigger/GlobalCaloTrigger/interface/L1GctJetCount.h"
 
 #include <iostream>
